Manage file browsing.

+++ alien4cloud-rest-api/src/main/java/org/alien4cloud/tosca/editor/EditorController.java
@@ -5,11 +5,15 @@ import java.io.InputStream;
 
 import javax.annotation.Resource;
 import javax.inject.Inject;
+import javax.validation.Valid;
 
 import org.alien4cloud.tosca.editor.model.EditionConcurrencyException;
 import org.alien4cloud.tosca.editor.operations.AbstractEditorOperation;
 import org.alien4cloud.tosca.editor.operations.UpdateFileOperation;
+import org.springframework.core.io.InputStreamResource;
+import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
@@ -30,8 +34,6 @@ import lombok.SneakyThrows;
 @RequestMapping({ "/rest/v2/editor", "/rest/latest/editor" })
 public class EditorController {
     @Inject
-    private TopologyDTOBuilder dtoBuilder;
-    @Inject
     private TopologyEditorService editorService;
     /** We use the artifact repository to store temporary files from the edition context. */
     @Resource
@@ -45,39 +47,71 @@ public class EditorController {
      */
     @SneakyThrows
     @ApiOperation(value = "Updates the deployment artifact of the node template.", notes = "The logged-in user must have the application manager role for this application. Application role required [ APPLICATION_MANAGER | APPLICATION_DEVOPS ]")
-    @RequestMapping(value = "/{topologyId}", method = RequestMethod.POST, produces = MediaType.APPLICATION_JSON_VALUE)
+    @RequestMapping(value = "/{topologyId}/execute", method = RequestMethod.POST, produces = MediaType.APPLICATION_JSON_VALUE)
     @PreAuthorize("isAuthenticated()")
-    public <T extends AbstractEditorOperation> RestResponse<TopologyDTO> execute(@PathVariable String topologyId, T operation) {
-        TopologyEditionContext context = editorService.execute(topologyId, operation);
-        return RestResponseBuilder.<TopologyDTO> builder().data(dtoBuilder.buildTopologyDTO(context)).build();
+    public RestResponse<TopologyDTO> execute(@PathVariable String topologyId, @RequestBody @Valid AbstractEditorOperation operation) {
+        TopologyDTO topologyDTO = editorService.execute(topologyId, operation);
+        return RestResponseBuilder.<TopologyDTO> builder().data(topologyDTO).build();
     }
 
     /**
      * Method exposed to REST to upload a file in an archive under edition.
      * 
      * @param topologyId The id of the topology/archive under edition.
-     * @param lastOperationId The id of the user last known operation (for optimistic locking edition).
+     * @param previousOperationId The id of the user last known operation (for optimistic locking edition).
      * @param path The path in which to save/override the file in the archive.
      * @param file The file to save in the archive.
      */
     @SneakyThrows
-    public void upload(@PathVariable String topologyId, String lastOperationId, String path, @RequestParam("file") MultipartFile file) throws IOException {
+    @PreAuthorize("isAuthenticated()")
+    @RequestMapping(value = "/{topologyId}/upload", method = RequestMethod.POST, produces = MediaType.APPLICATION_JSON_VALUE)
+    public RestResponse<TopologyDTO> upload(@PathVariable String topologyId, @RequestParam("previousOperationId") String previousOperationId,
+            @RequestParam("path") String path, @RequestParam(value = "file", required = false) MultipartFile file) throws IOException {
         // The controller saves the file in a temporary location and create a UpdateFileOperation to be sent in the edition context.
-        InputStream artifactStream = file.getInputStream();
         String artifactFileId = null;
-        try {
-            artifactFileId = artifactRepository.storeFile(artifactStream);
-            // FIXME return the topology context dto from the command execution ?
-        } finally {
-            Closeables.close(artifactStream, true);
+        if (file != null) {
+            InputStream artifactStream = file.getInputStream();
+            try {
+                artifactFileId = artifactRepository.storeFile(artifactStream);
+            } finally {
+                Closeables.close(artifactStream, true);
+            }
         }
         try {
-            editorService.execute(topologyId, new UpdateFileOperation(path, artifactFileId));
+            UpdateFileOperation updateFileOperation = new UpdateFileOperation(path, artifactFileId);
+            updateFileOperation.setPreviousOperationId(previousOperationId);
+            TopologyDTO topologyDTO = editorService.execute(topologyId, updateFileOperation);
+            return RestResponseBuilder.<TopologyDTO> builder().data(topologyDTO).build();
         } catch (EditionConcurrencyException e) {
             // Failed to perform the operation for concurrency issue, delete the temporary file.
-            artifactRepository.deleteFile(artifactFileId);
+            if (artifactFileId != null) {
+                artifactRepository.deleteFile(artifactFileId);
+            }
             throw e;
         }
+    }
+
+    /**
+     * Download a temporary file which is not yet commited (uploaded or modified through an operation).
+     * 
+     * @param topologyId The if of the topology.
+     * @param artifactId The id of the temporary artifact.
+     * @return The response entity with the input stream of the file.
+     */
+    @PreAuthorize("isAuthenticated()")
+    @RequestMapping(value = "/{topologyId}/file/{artifactId}", method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_VALUE)
+    public ResponseEntity<InputStreamResource> downloadTempFile(@PathVariable String topologyId, @PathVariable String artifactId) {
+        editorService.checkAuthorization(topologyId);
+
+        HttpHeaders headers = new HttpHeaders();
+        headers.add("Cache-Control", "no-cache, no-store, must-revalidate");
+        headers.add("Pragma", "no-cache");
+        headers.add("Expires", "0");
+
+        long length = artifactRepository.getFileLength(artifactId);
+
+        return ResponseEntity.ok().headers(headers).contentLength(length).contentType(MediaType.parseMediaType("application/octet-stream"))
+                .body(new InputStreamResource(artifactRepository.getFile(artifactId)));
     }
 
     /**
@@ -100,5 +134,4 @@ public class EditorController {
         // Call the service that will save and commit
 
     }
-
 }
